docker-app/backend/requirements.txt
#app
fastapi[all]
psycopg2-binary

#dev/tests
pytest
pytest-mock
mypy